SmartPotManualMode/requirements.txt
cherrypy==18.5.0
requests==2.25.1
paho-mqtt==1.5.1
cherrypy_cors==1.6


